+++ src/TransactionOutput.java
@@ -1,13 +1,13 @@
 public class TransactionOutput implements Cloneable{
-    private float value;
+    private double value;
     private String index = "";
     private String payeePublicKey = "";
 
-    public float getValue() {
+    public double getValue() {
         return value;
     }
 
-    public void setValue(float value) {
+    public void setValue(double value) {
         this.value = value;
     }
